+++ integration-tests/emulated/tests/bridges/bridge-hub-polkadot/src/lib.rs
@@ -44,6 +44,7 @@ pub use emulated_integration_tests_common::{
 	xcm_helpers::{xcm_transact_paid_execution, xcm_transact_unpaid_execution},
 	ASSETS_PALLET_ID, PROOF_SIZE_THRESHOLD, REF_TIME_THRESHOLD, XCM_V4,
 };
+pub use integration_tests_helpers::common::snowbridge::{MIN_ETHER_BALANCE, WETH};
 pub use kusama_polkadot_system_emulated_network::{
 	asset_hub_kusama_emulated_chain::{
 		genesis::ED as ASSET_HUB_KUSAMA_ED, AssetHubKusamaParaPallet as AssetHubKusamaPallet,
@@ -63,7 +64,11 @@ pub use kusama_polkadot_system_emulated_network::{
 		},
 		PenpalAssetOwner, PenpalBParaPallet as PenpalBPallet,
 	},
-	polkadot_emulated_chain::{genesis::ED as POLKADOT_ED, PolkadotRelayPallet as PolkadotPallet},
+	polkadot_emulated_chain::{
+		genesis::ED as POLKADOT_ED,
+		polkadot_runtime::xcm_config::UniversalLocation as PolkadotRelayUniversalLocation,
+		PolkadotRelayPallet as PolkadotPallet,
+	},
 	AssetHubKusamaPara as AssetHubKusama, AssetHubKusamaParaReceiver as AssetHubKusamaReceiver,
 	AssetHubKusamaParaSender as AssetHubKusamaSender, AssetHubPolkadotPara as AssetHubPolkadot,
 	AssetHubPolkadotParaReceiver as AssetHubPolkadotReceiver,