+++ relay/polkadot/constants/src/lib.rs
@@ -122,8 +122,13 @@ pub mod xcm {
 
 /// System Parachains.
 pub mod system_parachain {
-	use polkadot_primitives::Id;
+	use frame_support::parameter_types;
+	use polkadot_primitives::Id as ParaId;
 	use xcm_builder::IsChildSystemParachain;
+
+	parameter_types! {
+		pub AssetHubParaId: ParaId = ASSET_HUB_ID.into();
+	}
 
 	/// Asset Hub parachain ID.
 	pub const ASSET_HUB_ID: u32 = 1000;
@@ -137,7 +142,7 @@ pub mod system_parachain {
 	pub const BROKER_ID: u32 = 1005;
 
 	// System parachains from Polkadot point of view.
-	pub type SystemParachains = IsChildSystemParachain<Id>;
+	pub type SystemParachains = IsChildSystemParachain<ParaId>;
 
 	/// Coretime constants
 	pub mod coretime {